requirements.txt
yowsup2
pymongo